Renamed to offline/online for clarity
