+++ galasa-inttests-parent/dev.galasa.inttests/src/main/java/dev/galasa/inttests/compilation/simbank/local/offline/mvp/CompilationLocalJava08UbuntuMvp.java
@@ -3,7 +3,7 @@
  * 
  * (c) Copyright IBM Corp. 2021.
  */
-package dev.galasa.inttests.compilation.simbank.local.mvp;
+package dev.galasa.inttests.compilation.simbank.local.offline.mvp;
 
 import dev.galasa.Test;
 import dev.galasa.TestAreas;
