+++ galasa-inttests-parent/dev.galasa.inttests/src/main/java/dev/galasa/inttests/compilation/simbank/local/online/CompilationLocalJava08Ubuntu.java
@@ -3,7 +3,7 @@
  * 
  * (c) Copyright IBM Corp. 2021.
  */
-package dev.galasa.inttests.compilation.simbank.local;
+package dev.galasa.inttests.compilation.simbank.local.online;
 
 import dev.galasa.Test;
 import dev.galasa.TestAreas;
